Build scripts refactoring
