+++ Cargo.toml
@@ -1,5 +1,6 @@
 [workspace]
 members = [
+    "build-macros",
     "opengl-sys",
     "openal-sys",
     "libmirage-sys",
