+++ build-macros/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "build-macros"
+version = "1.0.0"
+authors = ["Marco Satti <marcosatti@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+bindgen = "0.53"
+serde = { version = "1.0" , features = ["derive"] }
+serde_json = "1.0"
