+++ build-macros/src/lib.rs
@@ -1,0 +1,5 @@
+mod external_build;
+mod external_check;
+
+pub use external_build::*;
+pub use external_check::*;
