+++ libpsx-rs/Cargo.toml
@@ -27,5 +27,4 @@ fixed = "0.5"
 typenum = "1.3"
 
 [build-dependencies]
-serde = { version = "1.0" , features = ["derive"] }
-serde_json = "1.0"
+build-macros = { path = "../build-macros" }
