+++ libpsx-rs/build.rs
@@ -1,7 +1,7 @@
-include!("../utilities/external_check.rs");
+use build_macros::external_check;
 
 fn main() {
-    external_check!("opengl", "opengl");
-    external_check!("openal", "openal");
-    external_check!("libmirage", "libmirage");
+    external_check("opengl", "opengl");
+    external_check("openal", "openal");
+    external_check("libmirage", "libmirage");
 }
