+++ psx-rs/Cargo.toml
@@ -16,5 +16,4 @@ libmirage-sys = { path = "../libmirage-sys" }
 libpsx-rs = { path = "../libpsx-rs" }
 
 [build-dependencies]
-serde = { version = "1.0" , features = ["derive"] }
-serde_json = "1.0"
+build-macros = { path = "../build-macros" }
